Updates documentation
Updates First Party Authentication guide to instruct whoever is reading it to set up Hooks and Rules.

Updates link to final version of the WildLogging SPWA (the version which has the admin stuff and authentication) in the README document and Third Party Authentication guide.

Also adds missing code to the Third Party Authentication guide.